porn91/plugins/v2ray/requirements.txt
aiohttp-socks==0.7.1
ping3==4.0.0
psutil==5.9.0